This commit brings pyArango significantly closer to PEP 8 complience.
Rationale:
The main reason for building a Python driver for an third-party API is,
or should be, to allow Pythonic access to the service that is clear and
understandable to a typical Python developer, and to simplify the process from having to use requests with custom JSON.

PEP 8 should be applied unless:
- There are project specific guidelines which take precedence. This is an
  open-source project and hence should not be creating its own
guidelines that are vastly different from the recommended guidlines.
- For backwards compatability. This is a new project, so there is no
  such requirement.
- Applying PEP 8 renders the code unreadable. This does not apply here,
  as the reason that I wrote this commit is precisely because I could
not read the code without PEP 8.
- The code predates the guidline. This does not apply.
- Compatability with older versions of Python. I have left some code
  intact for this reason (for example, explicitly extending the Object
class, which is redundant in Python3, has been left for Python 2
compatability). However, this library is likely to primarily be used by
Python3 developers, and a future pull-request will introduce Python3
style improvements and optimisations.

Many of the fixes in this commit also resolve *internal* inconsistencies
(for example, highly inconsistent doc-string usage, and differing names
for variables serving the equivalent function in different locations), which is a problem
regardless of the coding style adopted.

Key Changes:
- Splitting longer lines over multiple lines (PEP recomments 79
  character line length).
- Blank lines: From PEP, two lines between top-level definitions, one
  between methods. This had not been followed.
- Imports: some module-level imports were not at the top of files as
  required by PEP. They have been moved.
- Docstrings: Changed to triple quotes everywhere (PEP 257)
            : Placed the """ on separate lines. PEP 8 regards leaving
the trailing """ on a separate line as "most important". This had not been
followed.
- Whitespaces: Amazingly, almost every one of the "Pet Peeves" for
  whitespaces listed in PEP 8 were used. Theses are particularly
annoying and unnecessary. They have been removed.
- Naming conventions: I have not changed too many names because I am not
  yet familiar enough with the code to always have good names. However,
all variable and function names were CamelCase, which is incorrect. They
have been corrected.
                    : Single-letter variable names have been replaced
whereever possible.
                    : In particlar, the letter 'o' was used as a
variable name, which is highly inadvisable due to confusion with 0. O is
explicitly prohibited by PEP, but the same reasoning applies to o.
                    : I have replaced many abbreviations. Especially
cases where multiple DIFFERENT abbreviations were used for the same
variable name, leading to confusion.

Changes Still Required:
- Names that clash with reserved words have been incorrectly adapted
  (PEP 8 - Function and Method Arguments)
- Classes inhereting from Object don't need to do so explicitly, unless
there is a Python 2 reason for doing this.
- Function annotations may be helpful in certain cases.
- The actual content of the comments is currently entirely unhelpful and
docstrings in particular need to all be rewritten.